use BigDecimal instead of double

+++ balance-mng/balance-mng-facade/src/main/java/io/sofastack/balance/manage/facade/BalanceMngFacade.java
@@ -1,5 +1,7 @@
 package io.sofastack.balance.manage.facade;
 
+import java.math.BigDecimal;
+
 /**
  * @author yuanyuan
  * @since 2019/6/10
@@ -18,8 +20,9 @@ public interface BalanceMngFacade {
      * 返回用户余额
      *
      * @param userName 用户名
+     * @return
      */
-    Double queryBalance(String userName);
+    BigDecimal queryBalance(String userName);
 
     /**
      * 减少用户余额
@@ -27,5 +30,5 @@ public interface BalanceMngFacade {
      * @param userName 用户名
      * @param amount 减少数目
      */
-    void minusBalance(String userName, double amount);
+    void minusBalance(String userName, BigDecimal amount);
 }
