+++ stock-mng/src/main/java/io/sofastack/stockmng/mapper/StockMngMapper.java
@@ -3,6 +3,7 @@ package io.sofastack.stockmng.mapper;
 import io.sofastack.stockmng.model.ProductInfo;
 import org.apache.ibatis.annotations.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Mapper
@@ -15,7 +16,7 @@ public interface StockMngMapper {
     List<ProductInfo> query(@Param("userName") String userName);
 
     @Select("select price from stock_tb where product_code = #{productCode} and user_name = #{userName}")
-    Double queryProductPrice(@Param("productCode") String productCode, @Param("userName") String userName);
+    BigDecimal queryProductPrice(@Param("productCode") String productCode, @Param("userName") String userName);
 
     @Select("select sum(1) from stock_tb where user_name = #{userName}")
     Integer getStockRecordCountForUserName(@Param("userName") String userName);
@@ -24,7 +25,7 @@ public interface StockMngMapper {
         "insert into stock_tb (product_code, name, description, price, count, user_name) values (#{productCode}, "
             + "#{name}, #{description}, #{price}, #{count}, #{userName})")
     void insertStockRecord(@Param("productCode") String productCode, @Param("name") String name,
-                           @Param("description") String description, @Param("price") Double price,
+                           @Param("description") String description, @Param("price") BigDecimal price,
                            @Param("count") Integer count, @Param("userName") String userName);
 
     @Insert("insert into order_tb (user_name, product_code, count) values (#{userName}, #{productCode}, #{count})")
